refactor: context to hooks
