+++ apps/extension/src/ui/hooks/useCurrentSite.tsx
@@ -24,5 +24,3 @@ export const useCurrentSite = () => {
 
   return { id, url, title, favIconUrl, isLoading }
 }
-
-// export const [CurrentSiteProvider, useCurrentSite] = provideContext(useCurrentSiteProviderValue)
