+++ apps/extension/src/ui/hooks/usePopupNavOpenClose.ts
@@ -1,0 +1,9 @@
+import { useOpenCloseAtom } from "@talisman/hooks/useOpenClose"
+import { atom } from "recoil"
+
+const isNavOpenState = atom<boolean>({
+  key: "isNavOpenState",
+  default: false,
+})
+
+export const usePopupNavOpenClose = () => useOpenCloseAtom(isNavOpenState)
